Notification/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
sqlalchemy>=1.4.23
python-dotenv>=0.19.0
sendgrid>=6.8.0
aiosmtplib>=1.1.6
jinja2>=3.0.1
pytest>=6.2.5
httpx>=0.19.0